requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.14.0=pypi_0
aggdraw=1.3.15=pypi_0
aiohttp=3.7.4.post0=pypi_0
anyio=3.2.1=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.2=pypi_0
argon2-cffi=20.1.0=pypi_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
azure-core=1.19.0=pypi_0
azure-storage-blob=12.9.0=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pypi_0
black=21.9b0=pypi_0
blas=1.0=mkl
bleach=3.3.1=pypi_0
blis=0.7.4=pypi_0
bullet=2.2.0=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.7.5=hecd8cb5_1
cachetools=4.2.2=pypi_0
catalogue=2.0.4=pypi_0
certifi=2021.5.30=py38hecd8cb5_0
cffi=1.14.6=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
coloredlogs=15.0.1=pypi_0
configparser=5.0.2=pypi_0
cryptography=35.0.0=pypi_0
cycler=0.10.0=pypi_0
cymem=2.0.5=pypi_0
daemoniker=0.2.3=pypi_0
debugpy=1.3.0=pypi_0
decorator=5.0.9=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
en-core-web-md=3.1.0=pypi_0
en-core-web-trf=3.1.0=pypi_0
entrypoints=0.3=pypi_0
ffmpeg=4.3=h0a44026_0
filelock=3.7.0=pypi_0
freetype=2.10.4=ha233b18_0
fsspec=2021.8.1=pypi_0
future=0.18.2=pypi_0
getch=1.0=pypi_0
gettext=0.21.0=h7535e17_0
gitdb=4.0.7=pypi_0
gitpython=3.1.18=pypi_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
gridworld=0.1rc1=pypi_0
grpcio=1.40.0=pypi_0
gym=0.22.0=pypi_0
gym-notices=0.0.8=pypi_0
huggingface-hub=0.4.0=pypi_0
humanfriendly=9.2=pypi_0
icu=58.2=h0a44026_3
idna=2.10=pypi_0
iglu=0.2.2=pypi_0
importlib-metadata=5.0.0=pypi_0
inflection=0.5.1=pypi_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.3.0=hecd8cb5_3375
ipykernel=6.0.2=pypi_0
ipython=7.25.0=pypi_0
ipython-genutils=0.2.0=pypi_0
isodate=0.6.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=he5867d9_2
json5=0.9.6=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
jupyter-server=1.9.0=pypi_0
jupyterlab=3.0.16=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-server=2.6.1=pypi_0
kiwisolver=1.3.1=pypi_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.2.0=h9ed2024_0
libxml2=2.9.12=hcdb78fc_0
llvm-openmp=12.0.0=h0dcd299_1
llvmlite=0.38.1=pypi_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h23ab428_1
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.2.2=pypi_0
matplotlib-inline=0.1.2=pypi_0
minerl=0.4.2.1=pypi_0
minerl-patched=0.1.0=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.3.0=hecd8cb5_517
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.0=py38h4a7008c_2
mkl_random=1.2.2=py38hb2f4e1b_0
mpmath=1.2.1=pypi_0
msrest=0.6.21=pypi_0
multidict=5.1.0=pypi_0
murmurhash=1.0.5=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclassic=0.3.1=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.1.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pypi_0
nettle=3.7.3=h230ac6f_1
ninja=1.10.2=hf7b0b51_1
nltk=3.6.2=pypi_0
notebook=6.4.0=pypi_0
numba=0.55.2=pypi_0
numpy=1.21.5=pypi_0
numpy-base=1.20.3=py38he0bd621_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.5.64=pypi_0
openh264=2.1.0=hd9629dc_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1l=h9ed2024_0
packaging=21.0=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pathspec=0.9.0=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.6.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=pypi_0
pip=21.1.3=py38hecd8cb5_0
platformdirs=2.4.0=pypi_0
pluggy=1.0.0=pypi_0
preshed=3.0.5=pypi_0
prettytable=3.3.0=pypi_0
prometheus-client=0.11.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.19=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
py=1.10.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pydantic=1.8.2=pypi_0
pydeprecate=0.3.1=pypi_0
pyglet=1.5.21=pypi_0
pygments=2.9.0=pypi_0
pyopengl=3.1.5=pypi_0
pyparsing=2.4.7=pypi_0
pypi=2.1=pypi_0
pyro4=4.80=pypi_0
pyrsistent=0.18.0=pypi_0
pytest=6.2.5=pypi_0
python=3.8.10=h88f2d9e_7
python-dateutil=2.8.1=pypi_0
pytorch=1.11.0.dev20211007=py3.8_0
pytorch-lightning=1.7.2=pypi_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.1.0=pypi_0
readline=8.1=h9ed2024_0
regex=2021.7.6=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
requests-unixsocket=0.2.0=pypi_0
rsa=4.7.2=pypi_0
sacremoses=0.0.45=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.7.1=pypi_0
sentry-sdk=1.3.1=pypi_0
serpent=1.40=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=52.0.0=py38hecd8cb5_0
shapely=1.8.4=pypi_0
shortuuid=1.0.1=pypi_0
simple-term-menu=1.2.1=pypi_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
smart-open=5.1.0=pypi_0
smmap=4.0.0=pypi_0
sniffio=1.2.0=pypi_0
spacy=3.1.1=pypi_0
spacy-alignments=0.8.3=pypi_0
spacy-legacy=3.0.8=pypi_0
spacy-transformers=1.0.3=pypi_0
sqlite=3.36.0=hce871da_0
srsly=2.4.1=pypi_0
stable-baselines3=1.1.0=pypi_0
subprocess32=3.5.4=pypi_0
sympy=1.8=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.5.1=pypi_0
terminado=0.10.1=pypi_0
testpath=0.5.0=pypi_0
thinc=8.0.8=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.10=hb0a8c7a_0
tokenizers=0.10.3=pypi_0
toml=0.10.2=pypi_0
tomli=1.2.2=pypi_0
torch=1.9.1=pypi_0
torchaudio=0.11.0.dev20211007=py38_cpu
torchmetrics=0.9.3=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.12.0.dev20211007=py38_cpu
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.0.5=pypi_0
transformers=4.15.0=pypi_0
typer=0.3.2=pypi_0
typing=3.7.4.3=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.6=pypi_0
wandb=0.14.0=pypi_0
wasabi=0.8.2=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.1.0=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xmltodict=0.12.0=pypi_0
xz=5.2.5=h1de35cc_0
yarl=1.6.3=pypi_0
zipp=3.9.0=pypi_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
